CLASES-ROD

+++ Clase Phyton/Proyecto 2.py
@@ -1,5 +1,5 @@
 nombre = input("Por favor, dime tu nombre: ")
-ventas = int(input("Diga sus ventas totales del mes: "))
+ventas = float(input("Diga sus ventas totales del mes: "))
 
 comision = round(ventas * 13 / 100,2)
